v1.1
Updated CloudFormation template, so it points to the latest Lambda
function .zip file. Simplified price calculation logic so that it only
considers for calculation those resources that are tagged. Removed
rules that included untagged EC2 instances registered with a tagged
ELB, or an untagged ELB with tagged EC2 instances. Updated README.md to
show how to update the function using the CloudFormation template.